move create_mock_era5 to utils module
