+++ src/dmd_era5/era5_download/__init__.py
@@ -1,4 +1,3 @@
-from dmd_era5.era5_download.create_mock_era5 import create_mock_era5
 from dmd_era5.era5_download.era5_download import (
     config_parser,
     download_era5_data,
@@ -7,5 +6,4 @@ from dmd_era5.era5_download.era5_download import (
 __all__ = [
     "config_parser",
     "download_era5_data",
-    "create_mock_era5",
 ]
