+++ src/dmd_era5/era5_download/era5_download.py
@@ -7,9 +7,14 @@ import numpy as np
 import xarray as xr
 from pyprojroot import here
 
-from dmd_era5 import config_reader, log_and_print, setup_logger
-from dmd_era5.era5_download.create_mock_era5 import create_mock_era5
-from dmd_era5.utils import slice_era5_dataset, thin_era5_dataset
+from dmd_era5 import (
+    config_reader,
+    create_mock_era5,
+    log_and_print,
+    setup_logger,
+    slice_era5_dataset,
+    thin_era5_dataset,
+)
 
 config = config_reader("era5-download")
 logger = setup_logger("ERA5Download", "era5_download.log")
